+++ app/api/products/route.js
@@ -3,45 +3,57 @@ import connectMongoDB from "../../../libs/mongodb";
 import Product from "@/models/Product"
 
 export async function POST(req) {
-    try {
-        let product = await req.json();
-        await connectMongoDB();
-        const createdProduct = await Product.create(product);
+    const token = await getToken({ req });
 
-        return NextResponse.json(
-            {
-                message: "Product created successfully",
-                product: createdProduct
-            },
-            { status: 201 }
-        );
-    } catch (error) {
-        return NextResponse.json(
-            { message: "Error creating product", error: error.message },
-            { status: 500 }
-        );
+    if (token) {
+        try {
+            let product = await req.json();
+            await connectMongoDB();
+            const createdProduct = await Product.create(product);
+
+            return NextResponse.json(
+                {
+                    message: "Product created successfully",
+                    product: createdProduct
+                },
+                { status: 201 }
+            );
+        } catch (error) {
+            return NextResponse.json(
+                { message: "Error creating product", error: error.message },
+                { status: 500 }
+            );
+        }
+    } else {
+        return NextResponse.json({ message: "Not Authotized" }, { status: 401 });
     }
 }
 
 export async function PUT(req, response) {
-    try {
-        const products = await req.json();
+    const token = await getToken({ req });
 
-        await connectMongoDB();
+    if (token) {
+        try {
+            const products = await req.json();
 
-        const updatePromises = products.map(product =>
-            Product.updateOne(
-                { _id: product._id },
-                { $set: product }
-            )
-        );
+            await connectMongoDB();
 
-        await Promise.all(updatePromises);
+            const updatePromises = products.map(product =>
+                Product.updateOne(
+                    { _id: product._id },
+                    { $set: product }
+                )
+            );
 
-        return NextResponse.json({ message: "Products updated" }, { status: 200 });
-    } catch (error) {
-        console.error("Error updating products:", error);
-        return NextResponse.json({ error: "Failed to update products" }, { status: 500 });
+            await Promise.all(updatePromises);
+
+            return NextResponse.json({ message: "Products updated" }, { status: 200 });
+        } catch (error) {
+            console.error("Error updating products:", error);
+            return NextResponse.json({ error: "Failed to update products" }, { status: 500 });
+        }
+    } else {
+        return NextResponse.json({ message: "Not Authotized" }, { status: 401 });
     }
 }
 
@@ -52,28 +64,34 @@ export async function GET(req) {
 }
 
 export async function DELETE(req) {
-    try {
-        const id = req.nextUrl.searchParams.get("id");
+    const token = await getToken({ req });
 
-        if (!id) {
-            return NextResponse.json({ message: "Product ID is required" }, { status: 400 });
+    if (token) {
+        try {
+            const id = req.nextUrl.searchParams.get("id");
+
+            if (!id) {
+                return NextResponse.json({ message: "Product ID is required" }, { status: 400 });
+            }
+
+            await connectMongoDB();
+            const deletedProduct = await Product.findByIdAndDelete(id);
+
+            if (!deletedProduct) {
+                return NextResponse.json({ message: "Product not found" }, { status: 404 });
+            }
+
+            return NextResponse.json(
+                { message: "Product deleted", product: deletedProduct },
+                { status: 200 }
+            );
+        } catch (error) {
+            return NextResponse.json(
+                { message: "Error deleting product", error: error.message },
+                { status: 500 }
+            );
         }
-
-        await connectMongoDB();
-        const deletedProduct = await Product.findByIdAndDelete(id);
-
-        if (!deletedProduct) {
-            return NextResponse.json({ message: "Product not found" }, { status: 404 });
-        }
-
-        return NextResponse.json(
-            { message: "Product deleted", product: deletedProduct },
-            { status: 200 }
-        );
-    } catch (error) {
-        return NextResponse.json(
-            { message: "Error deleting product", error: error.message },
-            { status: 500 }
-        );
+    } else {
+        return NextResponse.json({ message: "Not Authotized" }, { status: 401 });
     }
 }
